requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0

# API Clients
stripe>=7.0.0
twilio>=8.0.0

# Database
psycopg2-binary>=2.9.0
redis>=5.0.0
sqlalchemy>=2.0.0

# Utilities
pyyaml>=6.0
jinja2>=3.1.0
schedule>=1.2.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0